Split up lynxkite_ops.py into smaller pieces.
